create Wiki Api folders and files
