--- auth/apps/frontend/src/components/markdown/wiki-generator.tsx
@@ -1,14 +1,0 @@
-import files, { metadata } from '@/components/markdown/wiki/home.mdx';
-
-export function WikiFiles() {
-	// const markdown_files = generateWikiMdx();
-	// const markdown_files = [{
-	//     id: "hey"
-	// }];
-
-	return (
-		<div>
-			{metadata.}
-		</div>
-	);
-}
