+++ auth/apps/wiki-api/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@tmlmobilidade/sae-plans-api",
+	"name": "@tmlmobilidade/wiki-api",
 	"version": "1.0.0",
 	"author": "",
 	"license": "ISC",
@@ -19,6 +19,7 @@
 		"@tmlmobilidade/interfaces": "20250708.1120.8",
 		"@tmlmobilidade/lib": "20250708.1120.8",
 		"@tmlmobilidade/utils": "20250708.1120.8",
+		"front-matter": "^4.0.2",
 		"pino-pretty": "13.0.0"
 	},
 	"devDependencies": {
