+++ auth/apps/wiki-api/src/endpoints/index.ts
@@ -1,2 +1,1 @@
-import '@/endpoints/plans/plans.routes';
-import '@/endpoints/validations/validations.routes';
+import '@/endpoints/wiki/wiki.routes';
