+++ auth/apps/wiki-api/src/endpoints/wiki/wiki.controller.ts
@@ -1,0 +1,19 @@
+/* * */
+
+import { type FastifyReply, type FastifyRequest } from '@tmlmobilidade/connectors';
+import { HttpStatus } from '@tmlmobilidade/lib';
+
+/* * */
+
+export class WikiController {
+	static async getAll(request: FastifyRequest, reply: FastifyReply) {
+		try {
+			reply.send(await markdownComponents.findMany({}, undefined, undefined, { created_at: -1 }));
+		}
+		catch (error) {
+			reply
+				.status(error.statusCode ?? HttpStatus.INTERNAL_SERVER_ERROR)
+				.send(error);
+		}
+	}
+}
