+++ auth/apps/wiki-api/src/utils/transform-mdx-to-html.ts
@@ -1,0 +1,23 @@
+/* * */
+
+interface ParsedMdxFile {
+	html: string
+	id: string
+	tags: string[]
+	title: string
+}
+
+export function transformMdxToHtml(mdxFilePath: string): ParsedMdxFile {
+	//
+
+	// .... trasfrom mdx to html ....
+
+	return {
+		html: '',
+		id: '',
+		tags: [],
+		title: '',
+	};
+
+	//
+}
